bcache: fix incorrect sysfs output value of strip size

+++ drivers/md/bcache/sysfs.c
@@ -209,7 +209,7 @@ SHOW(__bch_cached_dev)
 	sysfs_hprint(dirty_data,
 		     bcache_dev_sectors_dirty(&dc->disk) << 9);
 
-	sysfs_hprint(stripe_size,	dc->disk.stripe_size << 9);
+	sysfs_hprint(stripe_size, ((uint64_t)dc->disk.stripe_size) << 9);
 	var_printf(partial_stripes_expensive,	"%u");
 
 #if defined(CONFIG_BCACHE) || defined(CONFIG_BCACHE_MODULE)
